requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
jinja2>=3.1
pyyaml>=6.0
numpy>=1.26
faster-whisper>=1.0
llama-cpp-python>=0.2.84
reportlab>=4.2
PyQt5>=5.15
python-multipart>=0.0.9
